+++ tests/test_stringio.rs
@@ -1,4 +1,5 @@
 #[cfg(target_arch = "x86_64")]
+#[cfg(not(any(target_os = "windows", target_os = "macos")))]
 mod test_stringio {
     use runnel::medium::stringio::*;
     #[test]